requirements.txt
kafka-python==2.0.2
python-dotenv==1.0.1
feedparser==6.0.11
beautifulsoup4==4.12.3
html5lib==1.1
python-dateutil==2.9.0.post0
nltk==3.9.1
spacy==3.7.5
transformers==4.44.2
torch>=2.2.0
tokenizers>=0.15.2
pydantic>=2.7.0
